[FLINK-21180][python] Move the state module from 'pyflink.common' to 'pyflink.datastream'.
