+++ flink-python/pyflink/fn_execution/aggregate.py
@@ -18,7 +18,7 @@
 
 from apache_beam.coders import PickleCoder
 
-from pyflink.common.state import ListState, MapState
+from pyflink.datastream.state import ListState, MapState
 from pyflink.fn_execution.coders import from_proto
 from pyflink.fn_execution.operation_utils import is_built_in_function, load_aggregate_function
 from pyflink.fn_execution.state_impl import RemoteKeyedStateBackend
